+++ src/RugMessages.ts
@@ -1,4 +1,5 @@
-import { Attachment, emptyString, render } from "./SlackMessages";
+import { emptyString } from "./Common";
+import { Attachment, render } from "./SlackMessages";
 import { MessageMimeTypes, ResponseMessage } from "@atomist/rug/operations/Handlers";
 /**
  *